Use logical instead of list for dfs_order$dfs
Rather than a list of NULL values, and removing elements when the node has been
merged (slow), use a vector of logicals, and set them to `FALSE` when the
node has been merged.  This means that iterations must be over the subset of the
vector that is `TRUE`, but that is rarely calculated.